requirements.txt
Django==2.2
pytz==2022.2.1
sqlparse==0.4.2

djangorestframework~=3.13.1